Bug-fixing and ID-Validator for Requests
- Bug-fixing in functions „3.2.4 Edit a Video“ and „3.3.4 Edit an
Overlay“, so if the users wrote a „ ’ “ in a String (e.g. „… don’t … “)
the Database-querry failed before the fixing
- fixed bug in Cypher-Queries, so a Node-ID will be checked for it’s
corresponding label, e.g. for label „Videos“:
=> before: „START v=node(" + req.params.id + ") RETURN v“
=> after: „MATCH (v:Video) WHERE ID(v)=" + req.params.id + " RETURN v“
- same bug fixing for labels „Locations“ and „Overlays“
- Implemeted a ID-Validator for the functions „3.1.3 Retrieve a
Location“, „3.1.4 Edit a Location“, „3.1.5 Remove a Location“, „3.2.3
Retrieve a Video“, „3.2.4 Edit a Video“, „3.2.4 Remove a Video“, „3.3.3
Retrieve an Overlay“, „3.3.4 Edit an Overlay“, „3.3.5 Remove an
Overlay“ -> if a ID is not valid integer the HTTP-Status-Code is „406“
and the request failed